test: merge API promise test cases with service import
