--- src/transforms/v2-to-v3/__fixtures__/service-api-await.input.js
@@ -1,4 +1,0 @@
-import DynamoDB from "aws-sdk/clients/dynamodb";
-
-const client = new DynamoDB();
-const data = await client.listTables().promise();
