--- src/transforms/v2-to-v3/__fixtures__/service-api-await.output.js
@@ -1,4 +1,0 @@
-import { DynamoDB } from "@aws-sdk/client-dynamodb";
-
-const client = new DynamoDB();
-const data = await client.listTables();
