--- src/transforms/v2-to-v3/__fixtures__/service-api-promise-variable-declarator.output.js
@@ -1,4 +1,0 @@
-import { DynamoDB } from "@aws-sdk/client-dynamodb";
-
-const client = new DynamoDB();
-const listTablesPromise = client.listTables();
