--- src/transforms/v2-to-v3/__fixtures__/service-api-promise.input.js
@@ -1,9 +1,0 @@
-import DynamoDB from "aws-sdk/clients/dynamodb";
-
-const client = new DynamoDB();
-
-client
-  .listTables()
-  .promise()
-  .then((data) => console.log(data))
-  .catch((err) => console.log(err, err.stack));
