--- src/transforms/v2-to-v3/__fixtures__/service-api-promise.output.js
@@ -1,8 +1,0 @@
-import { DynamoDB } from "@aws-sdk/client-dynamodb";
-
-const client = new DynamoDB();
-
-client
-  .listTables()
-  .then((data) => console.log(data))
-  .catch((err) => console.log(err, err.stack));
